requirements.txt
streamlit
librosa
moviepy
mediapipe
openai-whisper
google-generativeai
opencv-python-headless
soundfile
numpy
